+++ packages/reusable-blocks/package.json
@@ -35,6 +35,7 @@
 		"@wordpress/i18n": "file:../i18n",
 		"@wordpress/icons": "file:../icons",
 		"@wordpress/notices": "file:../notices",
+		"@wordpress/url": "^2.19.0",
 		"lodash": "^4.17.19"
 	},
 	"publishConfig": {
